feat(Generators): ✨ Put color module behind feature and rename to colour
Add colour feauture and add to all
Add colour to feature test set in CI
Move color from data module to colour module
Update documentation to requie the feature
Update readme to refer to  colour
Update documentiaotn to refer to colour
Update code to refer to colour

Direct use of colour generators will require enabling the feature
