+++ src/colour/data.rs
