+++ src/data/mod.rs
@@ -1,4 +1,3 @@
 pub mod animal;
 pub mod beer;
-pub mod color;
 pub mod computer;
