+++ src/lib.rs
@@ -23,7 +23,8 @@ pub mod address;
 pub mod animal;
 pub mod beer;
 pub mod bool_rand;
-pub mod color;
+#[cfg(feature = "colour")]
+pub mod colour;
 #[cfg(feature = "company")]
 pub mod company;
 #[cfg(feature = "contact")]
